+++ kmad_web/services/spined.py
@@ -27,9 +27,21 @@ class SpinedService(object):
         out_file = os.path.join(SPINED_OUTPUT_DIR, tmp_name + '.spd')
         tmp_path = '/'.join(fasta_filename.split("/")[:-1])
         args = [self._path, tmp_path, tmp_name]
+        errlog_name = tmp_path + "_errlog"
         try:
-            subprocess.call(args)
-            os.remove(fasta_filename)
+            with open(errlog_name) as err:
+                subprocess.call(args, stderr=err)
+            # remove error log file if it's empty, otherwise raise an error
+            empty_errlog = os.stat(errlog_name).st_size == 0
+            if empty_errlog:
+                os.remove(empty_errlog)
+                os.remove(fasta_filename)
+            else:
+                e = "spine-d raised an error, check logfile: {}".format(
+                    errlog_name)
+                _log.error(e)
+                raise ServiceError(e)
+            # read output file
             if os.path.exists(out_file):
                 with open(out_file) as a:
                     data = a.read()
